// src/enums/token.enum.ts
export enum TokenEnum {
  PALAVRA_RESERVADA = "PALAVRA_RESERVADA",
  IDENTIFICADOR = "IDENTIFICADOR",
  NUMERO = "NUMERO",
  STRING = "STRING",
  OPERADOR = "OPERADOR",
  SEPARADOR = "SEPARADOR",
  SIMBOLO = "SIMBOLO",
  ESPACO = "ESPACO",
  DESCONHECIDO = "DESCONHECIDO",
  COMENTARIO = "COMENTARIO",
}